payment method specific data: moved objects to Entity\PaymentMethod #42
