+++ src/Transaction/PayTransaction.php
@@ -33,6 +33,7 @@
 namespace Wirecard\PaymentSdk\Transaction;
 
 use Wirecard\PaymentSdk\Entity\Money;
+use Wirecard\PaymentSdk\Entity\PaymentMethod\PayPalData;
 
 class PayTransaction implements Transaction
 {
